+++ src/modules/templatespecification/src/core/template_configuration.rs
@@ -1,7 +1,0 @@
-use crate::core::{file_list::FileList, template_specification::TemplateSpecification};
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct TemplateConfiguration {
-    pub template_specification: TemplateSpecification,
-    pub file_list: FileList,
-}